0827 goblin, mummy

+++ Assets/_MyProject/2. Scripts/GamePlay/Modules/Attackable.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class Attackable : MonoBehaviour, IAttackable
 {
     public int Damage { get; set; }
-    public int Range { get; set; }
+    public float Range { get; set; }
 
-    public void Initialize(int damage, int range)
+    public void Initialize(int damage, float range)
     {
         this.Damage = damage;
         this.Range = range;
